super_resolution_project/requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.21.0
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
